Added new  function to the RouteBuilder protocol and provided default implmentation that wraps original function.
