+++ Sources/Routing/Router.swift
@@ -7,14 +7,12 @@ public class Router {
     public final let base = Branch<Responder>(name: "", output: nil)
     
     /// A collection of all routes that have been registered with the Router
-    /// and their associated metadata.
-    /// Only the routes registered with the `register` function provided by MetadataRouteBuilder
-    /// will have user defined metadata
+    /// and their associated metadata, if any
     public var routeMetadata: [Route] = []
-
+    
     /// Init
     public init() {}
-
+    
     /// Register a given path. Use `*` for host OR method to define wildcards that will be matched
     /// if no concrete match exists.
     ///
@@ -40,7 +38,7 @@ public class Router {
         path: [String],
         metadata: [String: Any],
         responder: Responder
-    ) {
+        ) {
         let host = host ?? "*"
         let path = path.filter{ !$0.isEmpty }
         
@@ -55,7 +53,7 @@ public class Router {
     // caches static route resolutions
     private var _cache: [String: Responder?] = [:]
     private var _cacheLock = NSLock()
-
+    
     /// Removes all entries from this router's cache.
     ///
     public func flushCache() {
@@ -63,7 +61,7 @@ public class Router {
         _cache.removeAll()
         _cacheLock.unlock()
     }
-
+    
     /// Removes the cached Responder for a given Request.
     /// If there is no cached Responder, returns nil.
     ///
@@ -76,14 +74,14 @@ public class Router {
         _cacheLock.lock()
         let maybeCached = _cache.removeValue(forKey: request.routeKey)
         _cacheLock.unlock()
-
+        
         if let cached = maybeCached {
             return cached
         } else {
             return nil
         }
     }
-
+    
     /// Routes an incoming request
     /// the request will be populated with any found parameters (aka slugs).
     ///
@@ -92,18 +90,18 @@ public class Router {
         let key = request.routeKey
         
         // check the static route cache
-
+        
         _cacheLock.lock()
         let maybeCached = _cache[key]
         _cacheLock.unlock()
-
+        
         if let cached = maybeCached {
             return cached
         }
         
         let path = request.path()
         let result = base.fetch(path)
-
+        
         request.parameters = result?.slugs(for: path) ?? [:]
         
         // if there are no dynamic slugs, we can cache
@@ -133,7 +131,7 @@ extension Branch {
                 .flatMap({ $0.removingPercentEncoding })
                 .flatMap({ Parameters.string($0) })
                 else { return }
-
+            
             if let existing = slugs[key] {
                 var array = existing.array ?? [existing]
                 array.append(val)
@@ -177,4 +175,4 @@ extension Router {
 }
 
 extension Router: RouteBuilder {}
-extension Router: MetadataRouteBuilder {}
+
